requirements.txt
numpy==1.26.0
opencv-python==4.8.1.78
pytesseract==0.3.10
pyautogui==0.9.54
mss==9.0.1
tensorflow==2.16.1
pynput==1.7.6
psutil==5.9.6
win10toast==0.9
python-dotenv==1.0.0